+++ src/serialize_file.c
@@ -1,5 +1,5 @@
 
-
+#define R_NO_REMAP
 
 #include <R.h>
 #include <Rinternals.h>
@@ -15,7 +15,7 @@
 // Read a byte from the serialized stream
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 int read_byte_from_file(R_inpstream_t stream) {
-  error("read_byte_from_connection() never used for binary serializing");
+  Rf_error("read_byte_from_connection() never used for binary serializing");
   return 0;
 }
 
@@ -32,7 +32,7 @@ void read_bytes_from_file(R_inpstream_t stream, void *dst, int length) {
   
   // Sanity check that we read the requested number of bytes from the connection
   if (nread != length) {
-    error("read_bytes_from_file(). Expected %i bytes to be read, but actually read %ld", length, nread);
+    Rf_error("read_bytes_from_file(). Expected %i bytes to be read, but actually read %ld", length, nread);
   }
 }
 
@@ -41,7 +41,7 @@ void read_bytes_from_file(R_inpstream_t stream, void *dst, int length) {
 // Read a byte from the serialized stream
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 void write_byte_to_file(R_outpstream_t stream, int c) {
-  error("write_byte_to_connection() never used for binary serializing");
+  Rf_error("write_byte_to_connection() never used for binary serializing");
 }
 
 
@@ -62,7 +62,7 @@ SEXP unmarshall_file_(SEXP filename_) {
 
   FILE *fp = fopen(CHAR(STRING_ELT(filename_, 0)), "rb");
   if (fp == NULL) {
-    error("unmarshall_file(): Error opening '%s'", CHAR(STRING_ELT(filename_, 0)));
+    Rf_error("unmarshall_file(): Error opening '%s'", CHAR(STRING_ELT(filename_, 0)));
   }
   
   // Treat the data buffer as an input stream
@@ -94,7 +94,7 @@ SEXP marshall_file_(SEXP robj, SEXP filename_) {
   
   FILE *fp = fopen(CHAR(STRING_ELT(filename_, 0)), "wb");
   if (fp == NULL) {
-    error("unmarshall_file(): Error opening '%s'", CHAR(STRING_ELT(filename_, 0)));
+    Rf_error("unmarshall_file(): Error opening '%s'", CHAR(STRING_ELT(filename_, 0)));
   }
   
   // Create the output stream structure